Bugfix | Betrags Eingabe | Einzelumsatz direkt an Umsatz anlegen

+++ app/src/main/java/com/example/smartsave/EinzelumsatzEditActivity.kt
@@ -19,12 +19,14 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.smartsave.dataClasses.Einzelumsatz
 import com.example.smartsave.dataClasses.Kategorie
+import com.example.smartsave.dataClasses.Umsatz
 import com.example.smartsave.dataClasses.parseDate
 import com.example.smartsave.helpers.AlignedButton
 import com.example.smartsave.helpers.LabelledDatePickerButton
 import com.example.smartsave.helpers.labelledDropdownMenuKategory
 import com.example.smartsave.helpers.labelledDropdownMenuKategoryPreset
 import com.example.smartsave.helpers.LabelledInputField
+import com.example.smartsave.helpers.LabelledInputFieldBetrag
 import com.example.smartsave.helpers.MainColumn
 import com.example.smartsave.helpers.SmartSaveActivity
 import java.time.format.DateTimeFormatter
@@ -49,6 +51,9 @@ class EinzelumsatzEditActivity : SmartSaveActivity() {
         val einzelumsatz: Einzelumsatz? = if (mode == "editieren") {
             intent.extras?.getSerializable("einzelumsatz") as? Einzelumsatz
         } else { null }
+        val umsatz: Umsatz? = if (mode == "adden") {
+            intent.extras?.getSerializable("umsatz") as? Umsatz
+        } else { null }
 
         val formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy")
         var selectedDate by remember { mutableStateOf(einzelumsatz?.datum?.format(formatter) ?: "") }
@@ -56,6 +61,7 @@ class EinzelumsatzEditActivity : SmartSaveActivity() {
         var textBetrag by remember { mutableStateOf(einzelumsatz?.betrag?.toString() ?: "") }
         val kategorienListe by remember { kategorienListeState }
         var isError by remember { mutableStateOf(false) }
+        var isErrorBetrag by remember { mutableStateOf(false) }
         var selectedKategorie by remember { mutableStateOf(einzelumsatz?.kategorie) }
 
         MainColumn (
@@ -64,7 +70,10 @@ class EinzelumsatzEditActivity : SmartSaveActivity() {
             )
         {
             LabelledInputField(label = "Verwendungszweck*", value = textBezeichung, KeyboardOptions() ) {textBezeichung = it}
-            LabelledInputField(label = "Betrag*", value = textBetrag, KeyboardOptions(keyboardType = KeyboardType.Number) ) { textBetrag = it}
+            LabelledInputFieldBetrag(label = "Betrag*", value = textBetrag, KeyboardOptions(keyboardType = KeyboardType.Number) ) {
+                isErrorBetrag = false
+                textBetrag = it
+            }
             LabelledDatePickerButton(label = "Datum*",
                 selectedDate = selectedDate,
                 onDateSelected = { date -> selectedDate = date },
@@ -72,7 +81,7 @@ class EinzelumsatzEditActivity : SmartSaveActivity() {
                 maxdate = true
             )
 
-            if (mode == "anlegen") {
+            if (mode == "anlegen" || mode == "adden") {
                 selectedKategorie = labelledDropdownMenuKategory(label = "Kategorie", options = kategorienListe)
             } else if (mode == "editieren" && selectedKategorie != null) {
                 selectedKategorie = labelledDropdownMenuKategoryPreset(label = "Kategorie", options = kategorienListe, preset = selectedKategorie!!)
@@ -85,29 +94,50 @@ class EinzelumsatzEditActivity : SmartSaveActivity() {
                     style = MaterialTheme.typography.bodyLarge,
                     modifier = Modifier.padding(top = 4.dp)
                 )
+            } else if(isErrorBetrag) {
+                Text(
+                    text = "Bitte passenden Betrag angeben!",
+                    color = MaterialTheme.colorScheme.error,
+                    style = MaterialTheme.typography.bodyLarge,
+                    modifier = Modifier.padding(top = 4.dp)
+                )
             }
         }
 
         AlignedButton(alignment = Alignment.BottomStart, text = "Zurück") {finish()}
         AlignedButton(alignment = Alignment.BottomEnd, text = "Speichern"){
             isError = selectedDate.isEmpty() || textBezeichung.isEmpty() || textBetrag.isEmpty()
-            if(!isError){
+            if(!isError && !isErrorBetrag){
                 if (mode == "anlegen") {
                     val neweinzelumsatz = Einzelumsatz(textBezeichung,textBetrag.toDouble(), parseDate(selectedDate))
                     neweinzelumsatz.kategorie = selectedKategorie!!
-
                     db.insertEinzelumsatz(neweinzelumsatz)
+                    finish()
                 } else if (mode == "editieren" && einzelumsatz != null) {
                     einzelumsatz.verwendungsZweck = textBezeichung
                     einzelumsatz.betrag = textBetrag.toDouble()
                     einzelumsatz.datum = parseDate(selectedDate)
                     einzelumsatz.kategorie = selectedKategorie!!
-
                     db.removeEinzelumsatzZuweisung(einzelumsatz)
                     db.editEinzelumsatz(einzelumsatz)
+                    finish()
+                } else if (mode == "adden") {
+                    val neweinzelumsatz = Einzelumsatz(textBezeichung,textBetrag.toDouble(), parseDate(selectedDate))
+                    neweinzelumsatz.kategorie = selectedKategorie!!
+                    var restBetrag = umsatz!!.betrag
+                    for(einzelUmsatz in umsatz.einzelumsatzListe){
+                        restBetrag -= einzelUmsatz.betrag
+                    }
+                    if((umsatz.betrag > 0 && neweinzelumsatz.betrag > 0 && neweinzelumsatz.betrag <= restBetrag)
+                        || (umsatz.betrag < 0 && neweinzelumsatz.betrag < 0 && neweinzelumsatz.betrag >= restBetrag)) {
+                        neweinzelumsatz.id = db.insertEinzelumsatz(neweinzelumsatz)
+                        neweinzelumsatz.hasParentUmsatz = true
+                        db.addEinzelumsatzToUmsatz(umsatz.id, neweinzelumsatz)
+                        finish()
+                    } else {
+                        isErrorBetrag = true
+                    }
                 }
-
-                finish()
             }
         }
